mcp-rss/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "mcp-rss"
version = "1.0.0"
description = "MCP server for fetching, parsing, and monitoring RSS/Atom feeds"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "feedparser>=6.0.11",
    "beautifulsoup4>=4.12.0",
    "bleach>=6.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
]

[project.scripts]
mcp-rss = "src.server:main"

[tool.setuptools.packages.find]
where = ["."]
